+++ packages/api/src/context.ts
@@ -1,4 +1,5 @@
 import { auth } from "@ChannelGuide/auth";
+import prisma from "@ChannelGuide/db";
 import type { Context as HonoContext } from "hono";
 
 export type CreateContextOptions = {
@@ -10,7 +11,7 @@ export async function createContext({ context }: CreateContextOptions) {
     headers: context.req.raw.headers,
   });
   return {
-    auth: null,
+    prisma,
     session,
   };
 }